backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
tensorflow-cpu==2.17.0
numpy<2.0.0
pillow
